+++ src/Backup.ts
@@ -4,6 +4,7 @@ import joplin from "api";
 import * as path from "path";
 import backupLogging from "electron-log";
 import * as fs from "fs-extra";
+import * as os from "os";
 import { sevenZip } from "./sevenZip";
 import * as moment from "moment";
 import { helper } from "./helper";
@@ -16,6 +17,7 @@ class Backup {
   private msgDialog: any;
   private backupBasePath: string;
   private activeBackupPath: string;
+  private readmeOutputDirectory: string;
   private log: any;
   private logFile: string;
   private backupRetention: number;
@@ -28,6 +30,7 @@ class Backup {
   private compressionLevel: number;
   private singleJex: boolean;
   private createSubfolder: boolean;
+  private createSubfolderPerProfile: boolean;
   private backupSetName: string;
   private exportFormat: string;
   private execFinishCmd: string;
@@ -272,12 +275,10 @@ class Backup {
       );
     }
 
-    if (this.createSubfolder) {
-      this.log.verbose("append subFolder");
-      const orgBackupBasePath = this.backupBasePath;
-      this.backupBasePath = path.join(this.backupBasePath, "JoplinBackup");
+    const origBackupBasePath = this.backupBasePath;
+    const handleSubfolderCreation = async () => {
       if (
-        fs.existsSync(orgBackupBasePath) &&
+        fs.existsSync(origBackupBasePath) &&
         !fs.existsSync(this.backupBasePath)
       ) {
         try {
@@ -286,19 +287,81 @@ class Backup {
           await this.showError(i18n.__("msg.error.folderCreation", e.message));
         }
       }
+    };
+
+    if (this.createSubfolder) {
+      this.log.verbose("append subFolder");
+      this.backupBasePath = path.join(this.backupBasePath, "JoplinBackup");
+      await handleSubfolderCreation();
     }
 
-    if (path.normalize(profileDir) === this.backupBasePath) {
-      this.backupBasePath = null;
-      await this.showError(
-        i18n.__("msg.error.backupPathJoplinDir", path.normalize(profileDir))
+    // Set the README output directory before adding a subdirectory for the profile.
+    // This gives us one README for all backup subfolders.
+    this.readmeOutputDirectory = this.backupBasePath;
+
+    if (this.createSubfolderPerProfile) {
+      this.log.verbose("append profile subfolder");
+      // We assume that Joplin's profile structure is the following
+      //   rootProfileDir/
+      //   | profileDir/
+      //   | | [[profile content]]
+      // or, if using the default,
+      //   rootProfileDir/
+      //   | [[profile content]]
+      const profileRootDir = await joplin.settings.globalValue(
+        "rootProfileDir"
       );
+      const profileCurrentDir = await joplin.settings.globalValue("profileDir");
+
+      let profileName = path.basename(profileCurrentDir);
+      if (profileCurrentDir === profileRootDir) {
+        profileName = "default";
+      }
+
+      // Appending a -dev to the profile name prevents a devmode default Joplin
+      // profile from overwriting a non-devmode Joplin profile.
+      if ((await joplin.settings.globalValue("env")) === "dev") {
+        profileName += "-dev";
+      }
+
+      this.backupBasePath = path.join(this.backupBasePath, profileName);
+      await handleSubfolderCreation();
+    }
+
+    // Creating a backup can overwrite the backup directory. Thus,
+    // we mark several system and user directories as not-overwritable.
+    const systemDirectories = [
+      profileDir,
+      os.homedir(),
+
+      path.join(os.homedir(), "Desktop"),
+      path.join(os.homedir(), "Documents"),
+      path.join(os.homedir(), "Downloads"),
+      path.join(os.homedir(), "Pictures"),
+    ];
+
+    if (os.platform() === "win32") {
+      systemDirectories.push("C:\\Windows");
+    }
+
+    for (const systemDirectory of systemDirectories) {
+      if (helper.isSubdirectoryOrEqual(this.backupBasePath, systemDirectory)) {
+        const invalidBackupPath = this.backupBasePath;
+        this.backupBasePath = null;
+        await this.showError(
+          i18n.__("msg.error.backupPathContainsImportantDir", invalidBackupPath)
+        );
+        break;
+      }
     }
   }
 
   public async loadSettings() {
     this.log.verbose("loadSettings");
     this.createSubfolder = await joplin.settings.value("createSubfolder");
+    this.createSubfolderPerProfile = await joplin.settings.value(
+      "createSubfolderPerProfile"
+    );
     await this.loadBackupPath();
     this.backupRetention = await joplin.settings.value("backupRetention");
 
@@ -478,6 +541,7 @@ class Backup {
 
         const backupDst = await this.makeBackupSet();
 
+        await this.writeReadme(this.readmeOutputDirectory);
         await joplin.settings.setValue(
           "lastBackup",
           this.backupStartTime.getTime()
@@ -682,6 +746,16 @@ class Backup {
     } else {
       return false;
     }
+  }
+
+  private async writeReadme(backupFolder: string) {
+    const readmePath = path.join(backupFolder, "README.md");
+    this.log.info("writeReadme to", readmePath);
+    const readmeText = i18n.__(
+      "backupReadme",
+      this.backupStartTime.toLocaleString()
+    );
+    await fs.writeFile(readmePath, readmeText, "utf8");
   }
 
   private async backupNotebooks() {