+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## not released
 
+## v1.4.0 (2024-02-22)
+
+- Changes that are required for the Joplin default plugin
+- Renamed Plugin from `Simple Backup` to `Backup`
+- Add: Allow creating of subfolders for each profile
+
 ## v1.3.6 (2024-01-11)
 
 - Add: Screenshots / icon for [https://joplinapp.org/plugins/](https://joplinapp.org/plugins/)